+++ packages/sdk/src/swap/services/__tests__/ApiService.test.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { vi, describe, it, expect, beforeEach } from 'vitest';
 import { QuoteRequest } from '../../types';
-import { getQuote, getQuoteV2, getStatus, getStatusV2 } from '../ApiService';
+import { getQuoteV2, getStatusV2 } from '../ApiService';
 
 vi.mock('../../../../package.json', () => ({
   version: '1.0-test',
@@ -22,97 +22,6 @@ describe('ApiService', () => {
     destChain: 'Ethereum',
     destAsset: 'ETH',
   } satisfies QuoteRequest;
-
-  describe(getQuote, () => {
-    const mockedGet = vi.mocked(axios.get);
-    beforeEach(() => {
-      mockedGet.mockResolvedValueOnce({
-        data: {
-          id: 'string',
-          intermediateAmount: '1',
-          egressAmount: '2',
-        },
-      });
-    });
-
-    it('gets a quote', async () => {
-      const route = await getQuote('https://swapperoo.org', mockRoute, {});
-
-      expect(route).toMatchSnapshot();
-      expect(mockedGet.mock.lastCall).toMatchSnapshot();
-    });
-
-    it('gets a quote with a broker commission', async () => {
-      const route = await getQuote(
-        'https://swapperoo.org',
-        {
-          ...mockRoute,
-          brokerCommissionBps: 15,
-        },
-        {},
-      );
-
-      expect(route).toMatchSnapshot();
-      expect(mockedGet.mock.lastCall).toMatchSnapshot();
-    });
-
-    it('gets a quote with a broker commission and affiliate brokers', async () => {
-      const route = await getQuote(
-        'https://swapperoo.org',
-        {
-          ...mockRoute,
-          brokerCommissionBps: 15,
-          affiliateBrokers: [
-            { account: 'cFLdocJo3bjT7JbT7R46cA89QfvoitrKr9P3TsMcdkVWeeVLa', commissionBps: 10 },
-            { account: 'cFLdopvNB7LaiBbJoNdNC26e9Gc1FNJKFtvNZjAmXAAVnzCk4', commissionBps: 20 },
-          ],
-        },
-        {},
-      );
-
-      expect(route).toMatchSnapshot();
-      expect(mockedGet.mock.lastCall).toMatchSnapshot();
-    });
-
-    it('gets a quote for a vault swap', async () => {
-      const route = await getQuote(
-        'https://swapperoo.org',
-        {
-          ...mockRoute,
-          isVaultSwap: true,
-        },
-        {},
-      );
-
-      expect(route).toMatchSnapshot();
-      expect(mockedGet.mock.lastCall).toMatchSnapshot();
-    });
-
-    it('gets a quote with ccm params', async () => {
-      const route = await getQuote(
-        'https://swapperoo.org',
-        {
-          ...mockRoute,
-          ccmParams: {
-            gasBudget: '12345',
-            messageLengthBytes: 100,
-          },
-        },
-        {},
-      );
-
-      expect(route).toMatchSnapshot();
-      expect(mockedGet.mock.lastCall).toMatchSnapshot();
-    });
-
-    it('passes the signal to fetch', async () => {
-      await getQuote('https://swapperoo.org', mockRoute, {
-        signal: new AbortController().signal,
-      });
-
-      expect(mockedGet.mock.lastCall?.[1]?.signal).not.toBeUndefined();
-    });
-  });
 
   describe(getQuoteV2, () => {
     const mockedGet = vi.mocked(axios.get);
@@ -220,34 +129,6 @@ describe('ApiService', () => {
         {
           signal: new AbortController().signal,
         },
-      );
-
-      expect(mockedGet.mock.lastCall?.[1]?.signal).not.toBeUndefined();
-    });
-  });
-
-  describe(getStatus, () => {
-    it('forwards whatever response it gets from the swap service', async () => {
-      const mockedGet = vi.mocked(axios.get);
-      mockedGet.mockResolvedValueOnce({ data: 'hello darkness' });
-      mockedGet.mockResolvedValueOnce({ data: 'my old friend' });
-
-      const statusRequest = { id: 'the id' };
-
-      const status1 = await getStatus('https://swapperoo.org', statusRequest, {});
-      expect(status1).toBe('hello darkness');
-      const status2 = await getStatus('https://swapperoo.org', statusRequest, {});
-      expect(status2).toBe('my old friend');
-    });
-
-    it('passes the signal to axios', async () => {
-      const mockedGet = vi.mocked(axios.get);
-      mockedGet.mockResolvedValueOnce({ data: null });
-
-      await getStatus(
-        'https://swapperoo.org',
-        { id: '' },
-        { signal: new AbortController().signal },
       );
 
       expect(mockedGet.mock.lastCall?.[1]?.signal).not.toBeUndefined();