+++ packages/swap/src/handlers/__tests__/encodeVaultSwapData.test.ts
@@ -330,7 +330,7 @@ describe(encodeVaultSwapData, () => {
   });
 
   it('rejects if source asset is disabled', async () => {
-    env.FULLY_DISABLED_INTERNAL_ASSETS = ['Flip', 'Btc'];
+    env.FULLY_DISABLED_INTERNAL_ASSETS = new Set(['Flip', 'Btc']);
 
     await expect(
       encodeVaultSwapData({
@@ -354,7 +354,7 @@ describe(encodeVaultSwapData, () => {
   });
 
   it('rejects if destination asset is disabled', async () => {
-    env.FULLY_DISABLED_INTERNAL_ASSETS = ['Btc', 'Dot'];
+    env.FULLY_DISABLED_INTERNAL_ASSETS = new Set(['Btc', 'Dot']);
 
     await expect(
       encodeVaultSwapData({