+++ packages/swap/src/routes/__tests__/quote.test.ts
@@ -193,7 +193,7 @@ describe('server', () => {
     });
 
     it('rejects if source asset is disabled', async () => {
-      env.FULLY_DISABLED_INTERNAL_ASSETS = ['Flip', 'Btc'];
+      env.FULLY_DISABLED_INTERNAL_ASSETS = new Set(['Flip', 'Btc']);
 
       const params = new URLSearchParams({
         srcChain: 'Ethereum',
@@ -212,7 +212,7 @@ describe('server', () => {
     });
 
     it('rejects if destination asset is disabled', async () => {
-      env.FULLY_DISABLED_INTERNAL_ASSETS = ['Btc', 'Eth'];
+      env.FULLY_DISABLED_INTERNAL_ASSETS = new Set(['Btc', 'Eth']);
 
       const params = new URLSearchParams({
         srcChain: 'Ethereum',