+++ packages/swap/src/utils/__tests__/dataStructures.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, beforeEach, afterEach, expect, vi } from 'vitest';
-import { AsyncCacheMap, CacheMap } from '@/shared/dataStructures';
+import { AsyncCacheMap, Cache, CacheMap } from '@/shared/dataStructures';
 
 describe(CacheMap, () => {
   beforeEach(() => {
@@ -118,5 +118,53 @@ describe(AsyncCacheMap, () => {
 
     await expect(() => map.get('hello')).rejects.toThrow('nope');
     expect(await map.get('world')).toBe('hello world');
+  });
+});
+
+describe(Cache, () => {
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it('caches values for the TTL', async () => {
+    vi.useFakeTimers({ now: 0 });
+
+    let n = 0;
+
+    const cache = new Cache({
+      inc: {
+        async fetch() {
+          n += 1;
+          return n;
+        },
+        ttl: 10_000,
+      },
+    });
+
+    await expect(cache.read('inc')).resolves.toBe(1);
+    vi.advanceTimersByTime(5_000);
+    await expect(cache.read('inc')).resolves.toBe(1);
+    vi.advanceTimersByTime(5_000);
+    await expect(cache.read('inc')).resolves.toBe(1);
+    vi.advanceTimersByTime(1);
+    await expect(cache.read('inc')).resolves.toBe(2);
+  });
+
+  it('fetches fresh if the fetch fails', async () => {
+    let n = 0;
+
+    const cache = new Cache({
+      inc: {
+        async fetch() {
+          n += 1;
+          if (n === 1) throw new Error('try again');
+          return n;
+        },
+        ttl: 10_000,
+      },
+    });
+
+    await expect(cache.read('inc')).rejects.toThrow('try again');
+    await expect(cache.read('inc')).resolves.toBe(2);
   });
 });