+++ packages/swap/src/utils/env.ts
@@ -12,15 +12,15 @@ export const assertRouteEnabled = ({
   destAsset: ChainflipAsset;
 }) => {
   if (
-    env.FULLY_DISABLED_INTERNAL_ASSETS.includes(srcAsset) ||
-    env.DISABLED_DEPOSIT_INTERNAL_ASSETS.includes(srcAsset)
+    env.FULLY_DISABLED_INTERNAL_ASSETS.has(srcAsset) ||
+    env.DISABLED_DEPOSIT_INTERNAL_ASSETS.has(srcAsset)
   ) {
     throw ServiceError.unavailable(`Asset ${srcAsset} is disabled`);
   }
 
   if (
-    env.FULLY_DISABLED_INTERNAL_ASSETS.includes(destAsset) ||
-    env.DISABLED_DESTINATION_INTERNAL_ASSETS.includes(destAsset)
+    env.FULLY_DISABLED_INTERNAL_ASSETS.has(destAsset) ||
+    env.DISABLED_DESTINATION_INTERNAL_ASSETS.has(destAsset)
   ) {
     throw ServiceError.unavailable(`Asset ${destAsset} is disabled`);
   }